little configs refactor
